Sorting code almost complete.
Implemented the ability to find a pair of dual nodes based on a
"target", which is really just a pair of coordinates; that is, an
undirected edge. Have the ability to sort based on individual dual
nodes, but need to implement ability to sort based on *pairs* of dual
nodes, where only the shortest path of the pair matters.